+++ lib/src/route_parser.dart
@@ -4,25 +4,22 @@ import '../qlevar_router.dart';
 import 'types.dart';
 
 class QRouteInformationParser extends RouteInformationParser<MatchContext> {
-  final String _parent;
-
-  const QRouteInformationParser({@required String parent}) : _parent = parent;
-
+  const QRouteInformationParser();
   @override
   Future<MatchContext> parseRouteInformation(
       RouteInformation routeInformation) async {
     QR.log(
         // ignore: lines_longer_than_80_chars
-        'Searching for Route: ${routeInformation.location} for parent $_parent');
+        'Searching for Route: ${routeInformation.location}');
     if (routeInformation.location == null) {
       return SynchronousFuture(null);
     }
-    return QR.findMatch(routeInformation.location, parent: _parent);
+    return QR.findMatch(routeInformation.location);
   }
 
   @override
   RouteInformation restoreRouteInformation(MatchContext match) =>
-      RouteInformation(location:QR.currentRoute.fullPath);
+      RouteInformation(location: QR.currentRoute.fullPath);
 }
 
 class QRouteInformationProvider extends PlatformRouteInformationProvider {